+++ src/test/java/biz/gabrys/maven/plugins/directory/content/file/FileNameResolverTest.java
@@ -1,90 +1,118 @@
 package biz.gabrys.maven.plugins.directory.content.file;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
+import static org.mockito.Mockito.when;
+
 import java.io.File;
 
-import org.junit.Assert;
 import org.junit.Test;
-import org.mockito.Mockito;
 
 public final class FileNameResolverTest {
 
     @Test
     public void resolveName_fileWithoutExtension_returnFullName() {
         final FileNameResolver resolver = new FileNameResolver();
-        final File file = Mockito.mock(File.class);
-        Mockito.when(file.getName()).thenReturn("name");
-        Assert.assertEquals("File name.", "name", resolver.resolveName(file));
-        Mockito.verify(file).getName();
-        Mockito.verifyNoMoreInteractions(file);
+        final File file = mock(File.class);
+        when(file.getName()).thenReturn("name");
+
+        final String result = resolver.resolveName(file);
+
+        assertThat(result).isEqualTo("name");
+        verify(file).getName();
+        verifyNoMoreInteractions(file);
     }
 
     @Test
     public void resolveExtension_fileWithoutExtension_returnNull() {
         final FileNameResolver resolver = new FileNameResolver();
-        final File file = Mockito.mock(File.class);
-        Mockito.when(file.getName()).thenReturn("name");
-        Assert.assertEquals("File extension.", null, resolver.resolveExtension(file));
-        Mockito.verify(file).getName();
-        Mockito.verifyNoMoreInteractions(file);
+        final File file = mock(File.class);
+        when(file.getName()).thenReturn("name");
+
+        final String result = resolver.resolveExtension(file);
+
+        assertThat(result).isNull();
+        verify(file).getName();
+        verifyNoMoreInteractions(file);
     }
 
     @Test
     public void resolveName_fileWithExtension_returnNameWithoutExtension() {
         final FileNameResolver resolver = new FileNameResolver();
-        final File file = Mockito.mock(File.class);
-        Mockito.when(file.getName()).thenReturn("name.ext");
-        Assert.assertEquals("File name.", "name", resolver.resolveName(file));
-        Mockito.verify(file).getName();
-        Mockito.verifyNoMoreInteractions(file);
+        final File file = mock(File.class);
+        when(file.getName()).thenReturn("name.ext");
+
+        final String result = resolver.resolveName(file);
+
+        assertThat(result).isEqualTo("name");
+        verify(file).getName();
+        verifyNoMoreInteractions(file);
     }
 
     @Test
     public void resolveExtension_fileWithExtension_returnExtension() {
         final FileNameResolver resolver = new FileNameResolver();
-        final File file = Mockito.mock(File.class);
-        Mockito.when(file.getName()).thenReturn("name.ext");
-        Assert.assertEquals("File extension.", "ext", resolver.resolveExtension(file));
-        Mockito.verify(file).getName();
-        Mockito.verifyNoMoreInteractions(file);
+        final File file = mock(File.class);
+        when(file.getName()).thenReturn("name.ext");
+
+        final String result = resolver.resolveExtension(file);
+
+        assertThat(result).isEqualTo("ext");
+        verify(file).getName();
+        verifyNoMoreInteractions(file);
     }
 
     @Test
     public void resolveName_fileWithExtensionWithManyDots_returnNameWithoutExtension() {
         final FileNameResolver resolver = new FileNameResolver();
-        final File file = Mockito.mock(File.class);
-        Mockito.when(file.getName()).thenReturn("na.me.ext");
-        Assert.assertEquals("File name.", "na.me", resolver.resolveName(file));
-        Mockito.verify(file).getName();
-        Mockito.verifyNoMoreInteractions(file);
+        final File file = mock(File.class);
+        when(file.getName()).thenReturn("na.me.ext");
+
+        final String result = resolver.resolveName(file);
+
+        assertThat(result).isEqualTo("na.me");
+        verify(file).getName();
+        verifyNoMoreInteractions(file);
     }
 
     @Test
     public void resolveExtension_fileWithExtensionWithManyDots_returnExtension() {
         final FileNameResolver resolver = new FileNameResolver();
-        final File file = Mockito.mock(File.class);
-        Mockito.when(file.getName()).thenReturn("na.me.ext");
-        Assert.assertEquals("File extension.", "ext", resolver.resolveExtension(file));
-        Mockito.verify(file).getName();
-        Mockito.verifyNoMoreInteractions(file);
+        final File file = mock(File.class);
+        when(file.getName()).thenReturn("na.me.ext");
+
+        final String result = resolver.resolveExtension(file);
+
+        assertThat(result).isEqualTo("ext");
+        verify(file).getName();
+        verifyNoMoreInteractions(file);
     }
 
     @Test
     public void resolveName_fileWithDotAtTheEnd_returnNameWithoutLastDot() {
         final FileNameResolver resolver = new FileNameResolver();
-        final File file = Mockito.mock(File.class);
-        Mockito.when(file.getName()).thenReturn("name.");
-        Assert.assertEquals("File name.", "name", resolver.resolveName(file));
-        Mockito.verify(file).getName();
-        Mockito.verifyNoMoreInteractions(file);
+        final File file = mock(File.class);
+        when(file.getName()).thenReturn("name.");
+
+        final String result = resolver.resolveName(file);
+
+        assertThat(result).isEqualTo("name");
+        verify(file).getName();
+        verifyNoMoreInteractions(file);
     }
 
     @Test
     public void resolveExtension_fileWithDotAtTheEnd_returnEmptyText() {
         final FileNameResolver resolver = new FileNameResolver();
-        final File file = Mockito.mock(File.class);
-        Mockito.when(file.getName()).thenReturn("name.");
-        Assert.assertEquals("File extension.", "", resolver.resolveExtension(file));
-        Mockito.verify(file).getName();
-        Mockito.verifyNoMoreInteractions(file);
+        final File file = mock(File.class);
+        when(file.getName()).thenReturn("name.");
+
+        final String result = resolver.resolveExtension(file);
+
+        assertThat(result).isEmpty();
+        verify(file).getName();
+        verifyNoMoreInteractions(file);
     }
 }